refactor(testing): move signal test support into `ops::signal::testing`
`assert_builder_impls_match` lived in an inline `pub mod testing` inside
`window_builder.rs` and was reachable as `ops::signal::testing::*` only
because `ops/signal/mod.rs` glob-re-exports that module. `window_builder`
has no tests of its own and never called the helper; its sole consumer is
`cosine_window`.

Promote it to a real `ops/signal/testing/mod.rs`, declared directly rather
than glob-re-exported, so the namespace can hold more than one module's
helpers. Rename to `assert_sampling_window_builder_implementation`: the
old name leaned on the enclosing file to say what "builder" meant, which a
shared namespace no longer supplies.

Also widen `support::testing`'s gate to `any(test, feature = "testing")`,
matching this module. It previously resolved for in-crate unit tests only
via the dev-dependency self-reference unifying the feature.
